`Uplink` documentation and lib refactor
