+++ lib/src/main/java/io/bytebeam/uplink/Uplink.java
@@ -6,11 +6,14 @@ import android.content.Intent;
 import android.content.ServiceConnection;
 import android.os.*;
 import android.util.Log;
+import io.bytebeam.uplink.service.ActionSubscriber;
+import io.bytebeam.uplink.service.UplinkService;
+import io.bytebeam.uplink.service.UplinkTerminatedException;
 import io.bytebeam.uplink.types.ActionResponse;
 import io.bytebeam.uplink.types.UplinkAction;
 import io.bytebeam.uplink.types.UplinkPayload;
 
-import static io.bytebeam.uplink.UplinkService.*;
+import static io.bytebeam.uplink.service.UplinkService.*;
 
 enum ServiceState {
     UNINITIALIZED,
@@ -24,14 +27,25 @@ public class Uplink implements ServiceConnection {
     private final ServiceReadyCallback serviceReadyCallback;
     private Messenger serviceHandle;
     private ServiceState state = ServiceState.UNINITIALIZED;
+
     public ServiceState getState() {
         return state;
     }
 
+    /**
+     * Spawns an instance of the uplink.
+     *
+     * @param context              Current application context
+     * @param authConfig           authorization json configuration
+     * @param uplinkConfig         uplink toml configuration
+     * @param enableLogging        whether android logs should be reported as well
+     * @param serviceReadyCallback callback that will be invoked when the service is ready to be used
+     */
     public Uplink(
             Context context,
             String authConfig,
             String uplinkConfig,
+            boolean enableLogging,
             ServiceReadyCallback serviceReadyCallback
     ) {
         this.context = context;
@@ -39,7 +53,7 @@ public class Uplink implements ServiceConnection {
         Intent intent = new Intent(context, UplinkService.class);
         intent.putExtra(AUTH_CONFIG_KEY, authConfig);
         intent.putExtra(UPLINK_CONFIG_KEY, uplinkConfig);
-        Log.e(TAG, "triggering bind");
+        intent.putExtra(ENABLE_LOGGING_KEY, enableLogging);
         context.bindService(
                 intent,
                 this,
@@ -47,6 +61,11 @@ public class Uplink implements ServiceConnection {
         );
     }
 
+    /**
+     * Adds a subscriber that will be invoked when a new action is received.
+     *
+     * @throws UplinkTerminatedException if the uplink service has terminated for some reason
+     */
     public void subscribe(ActionSubscriber subscriber) throws UplinkTerminatedException {
         stateAssertion();
         Messenger messenger = new Messenger(
@@ -84,24 +103,44 @@ public class Uplink implements ServiceConnection {
         }
     }
 
+    /**
+     * Sends a payload to the uplink backend
+     *
+     * @throws UplinkTerminatedException if the uplink service has terminated for some reason
+     */
     public void sendData(UplinkPayload payload) throws UplinkTerminatedException {
         stateAssertion();
         callMethod(SEND_DATA, payload);
     }
 
+    /**
+     * Responds to an action that was received from the uplink backend
+     *
+     * @throws UplinkTerminatedException if the uplink service has terminated for some reason
+     */
     public void respondToAction(ActionResponse response) throws UplinkTerminatedException {
         stateAssertion();
         callMethod(SEND_DATA, response.toPayload());
     }
 
-    public void dispose() throws UplinkTerminatedException {
-        Log.e(TAG, "disposing uplink");
-        stateAssertion();
-        context.unbindService(this);
-        state = ServiceState.FINISHED;
+    /**
+     * To be called when the client is done using the service
+     * The uplink service will kill the process it was running in
+     * The instance must not be used after this method is called
+     */
+    public void dispose() {
+        try {
+            stateAssertion();
+            context.unbindService(this);
+            state = ServiceState.FINISHED;
+        } catch (UplinkTerminatedException e) {
+            Log.w(TAG, "Uplink service terminated before dispose was called");
+        }
     }
 
-    /** To be used for testing */
+    /**
+     * To be used for testing
+     */
     public void crash() throws UplinkTerminatedException {
         stateAssertion();
         callMethod(CRASH, null);
@@ -109,7 +148,6 @@ public class Uplink implements ServiceConnection {
 
     @Override
     public void onServiceConnected(ComponentName name, IBinder service) {
-        Log.e(TAG, "bind finished");
         state = ServiceState.CONNECTED;
         serviceHandle = new Messenger(service);
         serviceReadyCallback.uplinkReady();
@@ -118,7 +156,7 @@ public class Uplink implements ServiceConnection {
     @Override
     public void onServiceDisconnected(ComponentName name) {
         if (state != ServiceState.FINISHED) {
-           state = ServiceState.CRASHED;
+            state = ServiceState.CRASHED;
         }
     }
