+++ lib/src/main/java/io/bytebeam/uplink/service/UplinkTerminatedException.java
@@ -1,4 +1,4 @@
-package io.bytebeam.uplink;
+package io.bytebeam.uplink.service;
 
 /**
  * Thrown when the client notices that the remote
